--- backend/src/main/java/com/airassist/backend/exception/FontNotFoundException.java
@@ -1,7 +1,0 @@
-package com.airassist.backend.exception;
-
-public class FontNotFoundException extends RuntimeException {
-    public FontNotFoundException() {
-        super("Font not found");
-    }
-}
