+++ backend/src/main/java/com/airassist/backend/exceptionHandler/AirportApiExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.airassist.backend.exceptionHandler;
 
 import com.airassist.backend.controller.AirportApiController;
-import com.airassist.backend.exception.InvalidAirportDetailsException;
+import com.airassist.backend.exception.airport.InvalidAirportDetailsException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
