+++ backend/src/main/java/com/airassist/backend/service/PdfServiceImpl.java
@@ -1,6 +1,5 @@
 package com.airassist.backend.service;
 
-import com.airassist.backend.exception.FontNotFoundException;
 import com.openhtmltopdf.pdfboxout.PdfRendererBuilder;
 import lombok.AllArgsConstructor;
 import lombok.SneakyThrows;
@@ -9,12 +8,11 @@ import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
 
 import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
 import java.util.Map;
 
 @Service
 @AllArgsConstructor
-public class PdfServiceImplementation implements PdfService {
+public class PdfServiceImpl implements PdfService {
 
     private final TemplateEngine templateEngine;
 
@@ -28,16 +26,6 @@ public class PdfServiceImplementation implements PdfService {
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         PdfRendererBuilder builder = new PdfRendererBuilder();
         builder.withHtmlContent(htmlContent, null);
-
-
-        builder.useFont(() -> {
-            InputStream fontStream = getClass().getClassLoader().getResourceAsStream("fonts/DejaVuSans.ttf");
-
-            if (fontStream == null) {
-                throw new FontNotFoundException();
-            }
-            return fontStream;
-        }, "DejaVu Sans");
 
         builder.toStream(outputStream);
         builder.run();
