--- backend/src/main/resources/templates/ro_contract.html
@@ -1,132 +1,0 @@
-<!DOCTYPE html>
-<html lang="ro">
-<head>
-    <meta charset="UTF-8"/>
-    <title>Contract</title>
-    <style>
-        body {
-            font-family: 'DejaVu Sans', sans-serif;
-        }
-        .contract-box {
-            max-width: 800px;
-            margin: auto;
-            padding: 30px;
-            border: 1px solid #eee;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.15);
-        }
-        .contract-box table {
-            width: 100%;
-            line-height: inherit;
-            text-align: left;
-            border-collapse: collapse;
-        }
-        .contract-box table td {
-            padding: 8px;
-            vertical-align: top;
-            border: 1px solid #ddd;
-        }
-        .contract-box table tr.top td {
-            border: none;
-            padding-bottom: 20px;
-        }
-        .contract-box table tr.information td {
-            border: none;
-            padding-bottom: 40px;
-        }
-        .contract-box h2 {
-            margin: 0;
-            font-weight: normal;
-        }
-        .contract-box .section-title {
-            background: #eee;
-            font-weight: bold;
-            padding: 10px;
-            border: 1px solid #ddd;
-        }
-        .contract-box .no-border {
-            border: none;
-            padding: 0;
-        }
-    </style>
-</head>
-<body>
-<div class="contract-box">
-    <table>
-        <tr class="top">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td class="no-border">
-                            <h2>Acord Contractual</h2>
-                        </td>
-                        <td class="no-border" style="text-align:right;">
-                            ID Caz: <span th:text="${caseId}">123456789</span><br/>
-                            Data: <span th:text="${contractDate}">2025-07-31</span>
-                        </td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-
-        <tr class="information">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td colspan="2" class="section-title">Detalii Pasager</td>
-                    </tr>
-                    <tr>
-                        <td>Nume:</td>
-                        <td><span th:text="${passengerName}">John Doe</span></td>
-                    </tr>
-                    <tr>
-                        <td>Număr Pașaport:</td>
-                        <td><span th:text="${passportNumber}">A1234567</span></td>
-                    </tr>
-                    <tr>
-                        <td>Email:</td>
-                        <td><span th:text="${passengerEmail}">john.doe@example.com</span></td>
-                    </tr>
-                    <tr>
-                        <td>Telefon:</td>
-                        <td><span th:text="${passengerPhone}">+1 234 567 8900</span></td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-
-        <tr class="information">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td colspan="2" class="section-title">Detalii Rezervare</td>
-                    </tr>
-                    <tr>
-                        <td>Număr Rezervare:</td>
-                        <td><span th:text="${reservationNumber}">RES-2025-0001</span></td>
-                    </tr>
-                    <tr>
-                        <td>Date Călătorie:</td>
-                        <td><span th:text="${travelDates}">2025-08-15 până la 2025-08-20</span></td>
-                    </tr>
-                    <tr>
-                        <td>Destinație:</td>
-                        <td><span th:text="${destination}">Paris, Franța</span></td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-
-        <tr>
-            <td colspan="2" class="section-title">Termeni Contractuali</td>
-        </tr>
-        <tr>
-            <td colspan="2" style="border:1px solid #ddd; padding: 15px;">
-                <span>
-                    Prin semnarea acestui contract, pasagerul este de acord să respecte termenii și condițiile rezervării de călătorie, inclusiv politicile de anulare și cerințele de călătorie.
-                </span>
-            </td>
-        </tr>
-    </table>
-</div>
-</body>
-</html>
